feat: update main function and Application class #b423763
In the main function, the application now creates copies of files following the same directory structure within a .backup folder. Each copy is timestamped and appended to the end of the original filename in the format %Y-%m-%d_%H-%M-%S. The Application class no longer calls the update method upon initialization.